Protect against potential DDoS
The request body is used to compute the hash. As this is read into
memory, it would be possible for someone to send a very large request
payload that would exhaust the server's memory. To prevent this, we
limit the number of bytes we read from the request body stream to 1MB, a
default that can be overridden if desired.

Closes #1.